+++ GHEtool/Validation/artificial_neural_network.py
@@ -41,8 +41,8 @@ borefield.fluid_data = ConstantFluidData(0.568, 998, 4180, 1e-3)
 borefield.flow_data = ConstantFlowRate(mfr=0.2)
 borefield.pipe_data = DoubleUTube(1, 0.015, 0.02, 0.4, 0.05)
 borefield.calculation_setup(use_constant_Rb=False)
-borefield.set_max_avg_fluid_temperature(17)
-borefield.set_min_avg_fluid_temperature(3)
+borefield.set_max_fluid_temperature(17)
+borefield.set_min_fluid_temperature(3)
 hourly_load = HourlyGeothermalLoad()
 hourly_load.simulation_period = 20
 hourly_load.load_hourly_profile(FOLDER.joinpath("test\methods\hourly_data\office.csv"), header=True, separator=";",