+++ GHEtool/Validation/comparison_with_other_sizing_tools/test1a/test1a.py
@@ -46,8 +46,8 @@ def test_1a_6h():
     delta_t = max(load.max_peak_extraction, load.max_peak_injection) * 1000 / (fluid_data.cp() * flow_data.mfr())
 
     # set temperature bounds
-    borefield.set_max_avg_fluid_temperature(35 + delta_t / 2)
-    borefield.set_min_avg_fluid_temperature(0 - delta_t / 2)
+    borefield.set_max_fluid_temperature(35 + delta_t / 2)
+    borefield.set_min_fluid_temperature(0 - delta_t / 2)
 
     # according to L2
     L2_start = time.time()
@@ -80,8 +80,8 @@ def test_1a_6h():
     borefield.set_Rb(Rb_static)
 
     # set temperature bounds
-    borefield.set_max_avg_fluid_temperature(35 + delta_t / 2)
-    borefield.set_min_avg_fluid_temperature(0 - delta_t / 2)
+    borefield.set_max_fluid_temperature(35 + delta_t / 2)
+    borefield.set_min_fluid_temperature(0 - delta_t / 2)
 
     # load the hourly profile
     borefield.load = load
@@ -147,8 +147,8 @@ def test_1a_1h():
     delta_t = max(load.max_peak_extraction, load.max_peak_injection) * 1000 / (fluid_data.cp() * flow_data.mfr())
 
     # set temperature bounds
-    borefield.set_max_avg_fluid_temperature(35 + delta_t / 2)
-    borefield.set_min_avg_fluid_temperature(0 - delta_t / 2)
+    borefield.set_max_fluid_temperature(35 + delta_t / 2)
+    borefield.set_min_fluid_temperature(0 - delta_t / 2)
 
     # according to L2
     L2_start = time.time()
@@ -181,8 +181,8 @@ def test_1a_1h():
     borefield.set_Rb(Rb_static)
 
     # set temperature bounds
-    borefield.set_max_avg_fluid_temperature(35 + delta_t / 2)
-    borefield.set_min_avg_fluid_temperature(0 - delta_t / 2)
+    borefield.set_max_fluid_temperature(35 + delta_t / 2)
+    borefield.set_min_fluid_temperature(0 - delta_t / 2)
 
     # load the hourly profile
     borefield.load = load