+++ GHEtool/test/unit-tests/test_main_class.py
@@ -327,22 +327,26 @@ def test_set_pipe_params():
 
 def test_set_max_temp():
     borefield = Borefield()
-    borefield.set_max_avg_fluid_temperature(13)
+    borefield.set_max_fluid_temperature(13)
     assert borefield.Tf_max == 13
-    borefield.set_max_avg_fluid_temperature(14)
+    borefield.set_max_avg_fluid_temperature(15)
+    assert borefield.Tf_max == 15
+    borefield.set_max_fluid_temperature(14)
     assert borefield.Tf_max == 14
     with pytest.raises(ValueError):
-        borefield.set_max_avg_fluid_temperature(borefield.Tf_min - 1)
+        borefield.set_max_fluid_temperature(borefield.Tf_min - 1)
 
 
 def test_set_min_temp():
     borefield = Borefield()
-    borefield.set_min_avg_fluid_temperature(3)
+    borefield.set_min_fluid_temperature(3)
     assert borefield.Tf_min == 3
-    borefield.set_min_avg_fluid_temperature(4)
+    borefield.set_min_avg_fluid_temperature(5)
+    assert borefield.Tf_min == 5
+    borefield.set_min_fluid_temperature(4)
     assert borefield.Tf_min == 4
     with pytest.raises(ValueError):
-        borefield.set_min_avg_fluid_temperature(borefield.Tf_max + 1)
+        borefield.set_min_fluid_temperature(borefield.Tf_max + 1)
 
 
 def test_Tg():
@@ -448,7 +452,7 @@ def test_select_size():
     borefield.ground_data = data_ground_flux
     assert borefield._select_size(100, 20) == 100
     assert borefield._select_size(10, 80) == 80
-    borefield.set_max_avg_fluid_temperature(14)
+    borefield.set_max_fluid_temperature(14)
     with pytest.raises(UnsolvableDueToTemperatureGradient):
         borefield._select_size(10, 80)
 
@@ -494,7 +498,7 @@ def test_size_L3_value_errors():
 def test_size_L3(quadrant, result):
     borefield = Borefield()
     borefield.borefield = copy.deepcopy(borefield_gt)
-    borefield.set_max_avg_fluid_temperature(18)
+    borefield.set_max_fluid_temperature(18)
     borefield.load = MonthlyGeothermalLoadAbsolute(*load_case(2))
     borefield.ground_data = ground_data_constant
 
@@ -538,7 +542,7 @@ def test_size_L4():
     load.load_hourly_profile(FOLDER.joinpath("Examples/hourly_profile.csv"), col_injection=0, col_extraction=1)
     borefield.load = load
 
-    borefield.set_max_avg_fluid_temperature(25)
+    borefield.set_max_fluid_temperature(25)
     assert np.isclose(109.4742962707615, borefield.size_L4(100, quadrant_sizing=3))
     assert np.isclose(109.4742962707615, borefield.H)
     assert borefield.calculate_quadrant() == 3
@@ -1012,7 +1016,7 @@ def test_optimise_borefield_wrong_threshold_energy(monkeypatch):
 def test_calculate_quadrants_without_data():
     borefield = Borefield()
     borefield.borefield = copy.deepcopy(borefield_gt)
-    borefield.set_max_avg_fluid_temperature(18)
+    borefield.set_max_fluid_temperature(18)
     borefield.load = MonthlyGeothermalLoadAbsolute(*load_case(2))
     borefield.ground_data = ground_data_constant
     borefield.calculate_quadrant()
@@ -1172,8 +1176,8 @@ def test_optimise_load_borefield():
     load.load_hourly_profile(FOLDER.joinpath("Examples/hourly_profile.csv"))
     load.simulation_period = 10
     borefield = Borefield(load=load)
-    borefield.set_min_avg_fluid_temperature(2)
-    borefield.set_max_avg_fluid_temperature(17)
+    borefield.set_min_fluid_temperature(2)
+    borefield.set_max_fluid_temperature(17)
     borefield.borefield = gt.borefield.Borefield.rectangle_field(20, 4, 6, 6, 150, 1, 0.07)
     borefield.Rb = 0.1699
     ground_data = GroundFluxTemperature(2, 9.6, flux=0.07)
@@ -1288,7 +1292,7 @@ def test_with_titled_borefield():
     borefield.pipe_data = pipe_data
     borefield.fluid_data = fluid_data
     borefield.flow_data = ConstantFlowRate(mfr=0.2)
-    borefield.set_max_avg_fluid_temperature(17)
+    borefield.set_max_fluid_temperature(17)
 
     assert np.isclose(borefield.depth, 150 * math.cos(math.pi / 7) + 0.75)
     assert np.isclose(borefield.ground_data.calculate_Tg(borefield.depth, borefield.D), 12.157557845032045)